requirements.txt
# pip install -r requirements.txt

# --------- base --------- #
lightning-flash
icevision
lightning-flash[image]

# --------- hydra --------- #
hydra-core>=1.1.0
hydra-colorlog>=1.1.0
hydra-optuna-sweeper>=1.1.0

# --------- loggers --------- #
wandb~=0.12.16
scikit-learn~=1.1.1

# --------- linters --------- #
pre-commit      # hooks for applying linters on commit
black           # code formatting
isort           # import sorting
flake8          # code analysis
nbstripout      # remove output from jupyter notebooks

# --------- others --------- #
python-dotenv~=0.20.0
rich~=12.4.2
pytest          # tests
sh              # for running bash commands in some tests

omegaconf~=2.2.1
pillow~=8.4.0
numpy~=1.22.4
matplotlib~=3.5.2
seaborn~=0.11.2
torch~=1.10.2
bentoml~=1.0.0a7